Added a global type declaration.
