+++ types/NavTypes.d.ts
@@ -1,0 +1,8 @@
+type RefNames =
+| "Greeting"
+| "About"
+| "Education"
+| "Work"
+| "Skills"
+| "Projects"
+| "Contact";
